LayoutTests/fast/dom/Element/node-list-identity.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>

<ol name='name1' class='class2'/>
<ol name='name2' class='class1'/>
<p   name='name1' class='class1'/>

<script>
description("Test checking that node list fetched by various queries have proper object identity");

function areSame(expr) { shouldBeTrue(expr + ' === ' + expr) };

areSame('document.getElementsByTagName("ol")');
areSame('document.getElementsByName("name1")');
areSame('document.getElementsByClassName("class1")');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
