third_party/blink/web_tests/fast/dom/Element/node-list-identity-expected.txt

Test checking that node list fetched by various queries have proper object identity

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementsByTagName("ol") === document.getElementsByTagName("ol") is true
PASS document.getElementsByName("name1") === document.getElementsByName("name1") is true
PASS document.getElementsByClassName("class1") === document.getElementsByClassName("class1") is true
PASS successfullyParsed is true

TEST COMPLETE
